Dashboard: nest A4A Exclusive offers under agency supports, use /marketplace/exclusive-offers
Consistent with Tiers: the feature flag lives at supports.agency.exclusiveOffers
(dropping the separate supports.marketplace), the route is part of
createAgencyRoutes, and the URL is /marketplace/exclusive-offers. The Marketplace
sidebar group is gated by supports.agency.exclusiveOffers.
